refactor to seperate services from utils
