+++ student_app/lib/services/event_service.dart
