+++ student_app/lib/services/study_spot_service.dart
