+++ student_app/test/utils/event_service_test.dart
@@ -1,6 +1,6 @@
 import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:student_app/utils/event_service.dart';
+import 'package:student_app/services/event_service.dart';
 
 void main() {
   group('Event Collection Tests', () {
